+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2022-07-14
+
+Fixed an issue with non persistent classes.
+
+Added the ability to use a Locator rather than the static field to retrieve the class instance. In the future it would be good to find a way
+to remove the need for using a static instance field to define the Singleton.
+
 ## [1.0.6] - 2022-06-07
 
 Fixed the following: